Add the posibility to retry a wakup using a timestamp

+++ src/main/java/net/airvantage/sched/services/tech/RetryPolicyHelper.java
@@ -1,6 +1,11 @@
 package net.airvantage.sched.services.tech;
 
 import static net.airvantage.sched.quartz.job.JobResult.CallbackStatus.SUCCESS;
+
+import org.apache.commons.lang.Validate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import net.airvantage.sched.app.exceptions.AppException;
 import net.airvantage.sched.dao.JobWakeupDao;
 import net.airvantage.sched.model.JobState;
@@ -9,10 +14,6 @@ import net.airvantage.sched.quartz.job.JobResult;
 import net.airvantage.sched.quartz.job.JobResult.CallbackStatus;
 import net.airvantage.sched.services.JobSchedulingService;
 import net.airvantage.sched.services.JobStateService;
-
-import org.apache.commons.lang.Validate;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * A service to apply the retry policy.
@@ -68,18 +69,33 @@ public class RetryPolicyHelper {
         LOG.debug("handleResult : wakeup={}, result={}", wakeup, result);
 
         long requestedRetryDelay = result.getRetry();
+        long requestedRetryDate = result.getRetryDate();
 
-        if ((result.getStatus() == SUCCESS && requestedRetryDelay <= 0) || (wakeup.getRetryCount() >= MAX_RETRY_COUNT)) {
+        if ((result.getStatus() == SUCCESS && requestedRetryDelay <= 0 && requestedRetryDate <= 0)
+                || (wakeup.getRetryCount() >= MAX_RETRY_COUNT)) {
             // delete
             LOG.trace("handleResult deleting : wakeup={}, result={}", wakeup, result);
             jobWakeupDao.delete(wakeup.getId());
+
         } else {
             // rescheduling
-            int retryCount = wakeup.getRetryCount() + 1;
-            long retryDelay = Math.max(requestedRetryDelay, computeRetryDelay(retryCount));
 
+            // Increment the number of retry
+            int retryCount = wakeup.getRetryCount() + 1;
             wakeup.setRetryCount(retryCount);
-            wakeup.setWakeupTime(retryDelay + System.currentTimeMillis());
+
+            long now = System.currentTimeMillis();
+            if (requestedRetryDate > now) {
+                // A retry date is specified
+                long wakeupTime = Math.max(requestedRetryDate, roundNextSecond(now + computeRetryDelay(retryCount)));
+                wakeup.setWakeupTime(wakeupTime);
+
+            } else {
+                // A retry delay is specified
+                long retryDelay = Math.max(requestedRetryDelay, computeRetryDelay(retryCount));
+                wakeup.setWakeupTime(roundNextSecond(now + retryDelay));
+            }
+
             LOG.trace("handleResult rescheduling : wakeup={}, result={}", wakeup, result);
             jobWakeupDao.persist(wakeup);
         }
@@ -105,8 +121,11 @@ public class RetryPolicyHelper {
         // TODO check the retry date is before the next cron trigger fire time
     }
 
+    private long roundNextSecond(long timestamp) {
+        return (timestamp / 1000) * 1000 + 1000;
+    }
+
     private long computeRetryDelay(int retryCount) {
         return Math.min(MAX_RETRY_DELAY, Math.max(MIN_RETRY_DELAY, (long) Math.pow(2, retryCount)));
     }
-
 }
